Scientific equipment now works (almost)
things i should add/fix:descriptions for each equipment thing so its easier to know which one is for what type of anomaly, and maybe balance them a bit, also i think it works when unfolded, it shouldnt so that has to be changed somehow.  things that are now present: scientific equipment when placed ontop of an anomaly and activated will delete the anomaly, if its the right type of the anomaly, red looking sci equipment is for thermal anomalies, brown for gravitational ones, tesla rod looking one is for electric, green for acid, purple is for psy and some others that dont count as fire grav acid or electra.